+++ index.js
@@ -1,8 +1,0 @@
-'use strict';
-
-const { ApiListData, ApiListFilters } = require('./lib');
-
-module.exports = {
-	ApiListData,
-	ApiListFilters
-};