+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Removed
+- ApiListFilters was removed (**BREAKING CHANGE**)
+
+### Changed
+- API upgraded to v4 (`api-session` injected) (**BREAKING CHANGE**)
+
 ## [2.0.1] - 2019-09-16
 ### Fixed
 - Paging headers are now correctly parsed from string to number